finished arange_like, removed unit test

+++ tools/mo/openvino/tools/mo/front/mxnet/arange_like_ext.py
@@ -17,7 +17,7 @@ class ArangeLikeExt(FrontExtractorOp):
     def extract(cls, node: Node):
         attrs = get_mxnet_layer_attrs(node.symbol_dict)
         ArangeLikeOp.update_node_stat(node, {
-            'start': attrs.int('start', 0),
+            'start': attrs.float('start', 0),
             'repeat': attrs.int('repeat', 1),
             'step': attrs.float('step', 1),
             'axis': attrs.int('axis', None),
